test(utils): move utils tests into their src modules
Relocate tests/test_utils_{file,time,paths,directory,git}.rs into
`#[cfg(test)] mod tests` submodules inside the matching
src/utils/*.rs files, per the Rust Book test-organization guidance.

These are unit tests on individual utility helpers and belong next
to the code they cover; consolidating them removes five separate
integration-test crates and shortens compile time.
